Fix bullet manager time factor rounding.

+++ src/xrGame/Level_Bullet_Manager.h
@@ -218,7 +218,7 @@ protected:
 	bool					process_bullet		(
 								collide::rq_results& rq_storage,
 								SBullet& bullet,
-								u32 delta_time
+								float delta_time
 							);
 	void 		__stdcall	UpdateWorkload		();
 public:
